+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # CHANGELOG
 
 
+## v5.8.0 (2026-07-20)
+
+### Features
+
+- **NDX-817**: Make it possible to hot-reload novax programs on changes on the instance (#452)
+  ([#452](https://github.com/wandelbotsgmbh/wandelbots-nova/pull/452),
+  [`774006c`](https://github.com/wandelbotsgmbh/wandelbots-nova/commit/774006cdc6fc415f30c9867b0da96f4da03f568e))
+
+
+## v5.7.6 (2026-07-16)
+
+### Chores
+
+- Generate motion group models from api (#440)
+  ([#440](https://github.com/wandelbotsgmbh/wandelbots-nova/pull/440),
+  [`5eadb32`](https://github.com/wandelbotsgmbh/wandelbots-nova/commit/5eadb32467ed8a0c4511ce3b424f2b797d1c6155))
+
+
 ## v5.7.5 (2026-07-15)
 
 ### Bug Fixes